Enable sideloading for RPI
For the avocado-deploy-rpi SDK deploy hook to work, ssh must be available to
the SDK, and fwup must be available on the target.

Therefore, add nativesdk-openssh-ssh as an SDK dependency and fwup as an
application dependency.

Tested on a RPI CM5 via the command "avocado deploy -r dev -d <ip address>".

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>